Adjustments to get dev builds up. Still needs server deployment.

+++ build.sh
@@ -46,23 +46,19 @@ function usage {
 	cat >&2 <<DONE
 Usage: $0 [-p PLATFORMS] [-s DIR] [-v VERSION] [-c CHANNEL] [-d] [-t]
 Options
- -p PLATFORMS    *    build for platforms PLATFORMS (m=Mac, w=Windows, l=Linux)
- -s DIR               build symlinked to Zotero checkout DIR (implies -d)
- -v VERSION      *    use version VERSION (with leading "v")
- -c CHANNEL           use update channel CHANNEL
- -d                   don\'t package; only build binaries in staging/ directory
- -t                   build with debugging support
- -x XPI source   *    local, remote, or none
-
-(options marked with * are not optional)
+ -d DIR              build directory to build from (from build_xpi; cannot be used with -f)
+ -f FILE             ZIP file to build from (cannot be used with -d)
+ -t                  build with debugging support
+ -p PLATFORMS        build for platforms PLATFORMS (m=Mac, w=Windows, l=Linux)
+ -c CHANNEL          use update channel CHANNEL
+ -s                  don\'t package; only build binaries in staging/ directory
 DONE
 	exit 1
 }
 
 BUILD_DIR=`mktemp -d /tmp/tmp.XXXX`
 function cleanup {
-	#rm -rf $BUILD_DIR
-	echo "NOT DELETING $BUILD_DIR"
+	rm -rf $BUILD_DIR
 }
 trap cleanup EXIT
 
@@ -82,19 +78,28 @@ function seq () {
   fi
 }
 
+SOURCE_DIR=""
+ZIP_FILE=""
+BUILD_MAC=0
+BUILD_WIN32=0
+BUILD_LINUX=0
 PACKAGE=1
-while getopts "p:s:v:c:x:dt" opt; do
+DEVTOOLS=0
+while getopts "d:f:p:c:ts" opt; do
 	case $opt in
+		d)
+			SOURCE_DIR="$OPTARG"
+			;;
+		f)
+			ZIP_FILE="$OPTARG"
+			;;
 		p)
-			BUILD_MAC=0
-			BUILD_WIN32=0
-			BUILD_LINUX=0
 			for i in `seq 0 1 $((${#OPTARG}-1))`
 			do
 				case ${OPTARG:i:1} in
-					m) BUILD_MAC=1;GECKO_VERSION="40.0";GECKO_SHORT_VERSION="40.0";;
-					w) BUILD_WIN32=1;GECKO_VERSION="40.0";GECKO_SHORT_VERSION="40.0";;
-					l) BUILD_LINUX=1;GECKO_VERSION="39.0";GECKO_SHORT_VERSION="39.0";;
+					m) BUILD_MAC=1;;
+					w) BUILD_WIN32=1;;
+					l) BUILD_LINUX=1;;
 					*)
 						echo "$0: Invalid platform option ${OPTARG:i:1}"
 						usage
@@ -102,24 +107,14 @@ while getopts "p:s:v:c:x:dt" opt; do
 				esac
 			done
 			;;
-		s)
-			SYMLINK_DIR="$OPTARG"
-			PACKAGE=0
-			;;
-		v)
-			VERSION="$OPTARG"
-			;;
 		c)
 			UPDATE_CHANNEL="$OPTARG"
 			;;
-		x)
-			XPI_SOURCE="$OPTARG"
-			;;
-		d)
-			PACKAGE=0
-			;;
 		t)
 			DEVTOOLS=1
+			;;
+		s)
+			PACKAGE=0
 			;;
 		*)
 			usage
@@ -128,42 +123,40 @@ while getopts "p:s:v:c:x:dt" opt; do
 	shift $((OPTIND-1)); OPTIND=1
 done
 
-if [ ${BUILD_LINUX} -eq 0 -a ${BUILD_MAC} -eq 0 -a ${BUILD_WIN32} -eq 0 ]; then
-    echo "Must set the platform (-p) option"
-    usage
-fi
 
-# Must set the version (-v) option
-if [ "${VERSION}" == "" ]; then
-    echo "Must set the version (-v) option"
-    usage
-fi
+# My Last Param Variables
+#
+#			BUILD_MAC=0
+#			BUILD_WIN32=0
+#			BUILD_LINUX=0
+#					m) BUILD_MAC=1;GECKO_VERSION="40.0";GECKO_SHORT_VERSION="40.0";;
+#					w) BUILD_WIN32=1;GECKO_VERSION="40.0";GECKO_SHORT_VERSION="40.0";;
+#					l) BUILD_LINUX=1;GECKO_VERSION="39.0";GECKO_SHORT_VERSION="39.0";;
+#			SYMLINK_DIR="$OPTARG"
+#			PACKAGE=0
+#			VERSION="$OPTARG"
+#			UPDATE_CHANNEL="$OPTARG"
+#			XPI_SOURCE="$OPTARG"
+#			PACKAGE=0
+#			DEVTOOLS=1
 
-if [ "${XPI_SOURCE}" != "local" -a "${XPI_SOURCE}" != "remote" -a "${XPI_SOURCE}" != "none" ]; then
-    echo "(${XPI_SOURCE})"
-    usage
-fi 
 
-# Not sure what this protects against.
-if [ ! -z $1 ]; then
-    echo FOUR
+# Require source dir or ZIP file
+if [[ -z "$SOURCE_DIR" ]] && [[ -z "$ZIP_FILE" ]]; then
+	usage
+elif [[ -n "$SOURCE_DIR" ]] && [[ -n "$ZIP_FILE" ]]; then
 	usage
 fi
 
-VERSION=$(echo "${VERSION}" | sed -e "s/v\(.*\)/\1/")
+# Require at least one platform
+if [[ $BUILD_MAC == 0 ]] && [[ $BUILD_WIN32 == 0 ]] && [[ $BUILD_LINUX == 0 ]]; then
+	usage
+fi
 
-echo "BUILD_LINUX=${BUILD_LINUX}"
-echo "BUILD_MAC=${BUILD_MAC}"
-echo "BUILD_WIN32=${BUILD_WIN32}"
-echo "XPI_SOURCE=${XPI_SOURCE}"
-echo "VERSION=${VERSION}"
-
-. grab_xpis.sh "${BUILD_LINUX}${BUILD_MAC}${BUILD_WIN32}" "${XPI_SOURCE}" "$CALLDIR" "$BUILD_DIR"
-
-BUILD_ID=`date +%Y%m%d`
+BUILD_ID=`date +%Y%m%d%H%M%S`
 
 shopt -s extglob
-mkdir -p "$BUILD_DIR/jurism"
+mkdir -p "$BUILD_DIR/zotero"
 rm -rf "$STAGE_DIR"
 mkdir "$STAGE_DIR"
 rm -rf "$DIST_DIR"
@@ -172,65 +165,73 @@ mkdir "$DIST_DIR"
 # Save build id, which is needed for updates manifest
 echo $BUILD_ID > "$DIST_DIR/build_id"
 
-# Add devtools manifest and pref
-if [ $DEVTOOLS -eq 1 ]; then
-	cat "$CALLDIR/assets/devtools.manifest" >> "$BUILD_DIR/jurism/chrome.manifest"
-	echo 'pref("devtools.debugger.remote-enabled", true);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-	echo 'pref("devtools.debugger.remote-port", 6100);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-	echo 'pref("devtools.debugger.prompt-connection", false);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-fi
-
 if [ -z "$UPDATE_CHANNEL" ]; then UPDATE_CHANNEL="default"; fi
 
-cd "$BUILD_DIR/jurism"
+if [ -n "$ZIP_FILE" ]; then
+	ZIP_FILE="`abspath $ZIP_FILE`"
+	echo "Building from $ZIP_FILE"
+	unzip -q $ZIP_FILE -d "$BUILD_DIR/zotero"
+else
+	# TODO: Could probably just mv instead, at least if these repos are merged
+	rsync -a "$SOURCE_DIR/" "$BUILD_DIR/zotero/"
+fi
 
-# Upstream Zotero code to extract version number from install.rdf
-# omitted here.
+cd "$BUILD_DIR/zotero"
 
+VERSION=`perl -ne 'print and last if s/.*<em:version>(.*)<\/em:version>.*/\1/;' install.rdf`
+if [ -z "$VERSION" ]; then
+	echo "Version number not found in install.rdf"
+	exit 1
+fi
 rm install.rdf
 
 echo
 echo "Version: $VERSION"
 
+#
+# Restore this eventually, for the additional plugins
+# . grab_xpis.sh "${BUILD_LINUX}${BUILD_MAC}${BUILD_WIN32}" "${XPI_SOURCE}" "$CALLDIR" "$BUILD_DIR"
+#
+
 # Delete Mozilla signing info if present
 rm -rf META-INF
 
 # Copy branding
-cp -R "$CALLDIR/assets/branding" "$BUILD_DIR/jurism/chrome/branding"
+cp -R "$CALLDIR/assets/branding" "$BUILD_DIR/zotero/chrome/branding"
 
 # Add to chrome manifest
-echo "" >> "$BUILD_DIR/jurism/chrome.manifest"
-cat "$CALLDIR/assets/chrome.manifest" >> "$BUILD_DIR/jurism/chrome.manifest"
+echo "" >> "$BUILD_DIR/zotero/chrome.manifest"
+cat "$CALLDIR/assets/chrome.manifest" >> "$BUILD_DIR/zotero/chrome.manifest"
 
 # Copy Error Console files
-cp "$CALLDIR/assets/console/jsconsole-clhandler.js" "$BUILD_DIR/jurism/components/"
-echo >> "$BUILD_DIR/jurism/chrome.manifest"
-cat "$CALLDIR/assets/console/jsconsole-clhandler.manifest" >> "$BUILD_DIR/jurism/chrome.manifest"
-cp -R "$CALLDIR/assets/console/content" "$BUILD_DIR/jurism/chrome/console"
-cp -R "$CALLDIR/assets/console/skin/osx" "$BUILD_DIR/jurism/chrome/console/skin"
-cp -R "$CALLDIR/assets/console/locale/en-US" "$BUILD_DIR/jurism/chrome/console/locale"
-cat "$CALLDIR/assets/console/jsconsole.manifest" >> "$BUILD_DIR/jurism/chrome.manifest"
+cp "$CALLDIR/assets/console/jsconsole-clhandler.js" "$BUILD_DIR/zotero/components/"
+echo >> "$BUILD_DIR/zotero/chrome.manifest"
+cat "$CALLDIR/assets/console/jsconsole-clhandler.manifest" >> "$BUILD_DIR/zotero/chrome.manifest"
+cp -R "$CALLDIR/assets/console/content" "$BUILD_DIR/zotero/chrome/console"
+cp -R "$CALLDIR/assets/console/skin/osx" "$BUILD_DIR/zotero/chrome/console/skin"
+cp -R "$CALLDIR/assets/console/locale/en-US" "$BUILD_DIR/zotero/chrome/console/locale"
+cat "$CALLDIR/assets/console/jsconsole.manifest" >> "$BUILD_DIR/zotero/chrome.manifest"
 
 # Delete files that shouldn't be distributed
-${GFIND} "$BUILD_DIR/jurism/chrome" -name .DS_Store -exec rm -f {} \;
+find "$BUILD_DIR/zotero/chrome" -name .DS_Store -exec rm -f {} \;
 
 # Zip chrome into JAR
-cd "$BUILD_DIR/jurism"
+cd "$BUILD_DIR/zotero"
 zip -r -q jurism.jar chrome deleted.txt resource styles.zip translators.index translators.zip styles translators.json translators
 rm -rf "chrome/"* install.rdf deleted.txt resource styles.zip translators.index translators.zip styles translators.json translators
 
 # Copy updater.ini
-cp "$CALLDIR/assets/updater.ini" "$BUILD_DIR/jurism"
+cp "$CALLDIR/assets/updater.ini" "$BUILD_DIR/zotero"
 
 # Adjust chrome.manifest
-perl -pi -e 's^(chrome|resource)/^jar:jurism.jar\!/$1/^g' "$BUILD_DIR/jurism/chrome.manifest"
+perl -pi -e 's^(chrome|resource)/^jar:jurism.jar\!/$1/^g' "$BUILD_DIR/zotero/chrome.manifest"
 
 # Adjust connector pref
-perl -pi -e 's/pref\("extensions\.zotero\.httpServer\.enabled", false\);/pref("extensions.zotero.httpServer.enabled", true);/g' "$BUILD_DIR/jurism/defaults/preferences/zotero.js"
-perl -pi -e 's/pref\("extensions\.zotero\.connector\.enabled", false\);/pref("extensions.zotero.connector.enabled", true);/g' "$BUILD_DIR/jurism/defaults/preferences/zotero.js"
+perl -pi -e 's/pref\("extensions\.zotero\.httpServer\.enabled", false\);/pref("extensions.zotero.httpServer.enabled", true);/g' "$BUILD_DIR/zotero/defaults/preferences/zotero.js"
+perl -pi -e 's/pref\("extensions\.zotero\.connector\.enabled", false\);/pref("extensions.zotero.connector.enabled", true);/g' "$BUILD_DIR/zotero/defaults/preferences/zotero.js"
 
 # Copy icons
-cp -r "$CALLDIR/assets/icons" "$BUILD_DIR/jurism/chrome/icons"
+cp -r "$CALLDIR/assets/icons" "$BUILD_DIR/zotero/chrome/icons"
 
 # Copy application.ini and modify
 cp "$CALLDIR/assets/application.ini" "$BUILD_DIR/application.ini"
@@ -238,25 +239,24 @@ perl -pi -e "s/\{\{VERSION}}/$VERSION/" "$BUILD_DIR/application.ini"
 perl -pi -e "s/\{\{BUILDID}}/$BUILD_ID/" "$BUILD_DIR/application.ini"
 
 # Copy prefs.js and modify
-cp "$CALLDIR/assets/prefs.js" "$BUILD_DIR/jurism/defaults/preferences"
-perl -pi -e 's/pref\("app\.update\.channel", "[^"]*"\);/pref\("app\.update\.channel", "'"$UPDATE_CHANNEL"'");/' "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-perl -pi -e 's/%GECKO_VERSION%/'"$GECKO_VERSION"'/g' "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+cp "$CALLDIR/assets/prefs.js" "$BUILD_DIR/zotero/defaults/preferences"
+perl -pi -e 's/pref\("app\.update\.channel", "[^"]*"\);/pref\("app\.update\.channel", "'"$UPDATE_CHANNEL"'");/' "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 
 # Add devtools manifest and pref
 if [ $DEVTOOLS -eq 1 ]; then
-	cat "$CALLDIR/assets/devtools.manifest" >> "$BUILD_DIR/jurism/chrome.manifest"
-	echo 'pref("devtools.debugger.remote-enabled", true);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-	echo 'pref("devtools.debugger.remote-port", 6100);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
-	echo 'pref("devtools.debugger.prompt-connection", false);' >> "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+	cat "$CALLDIR/assets/devtools.manifest" >> "$BUILD_DIR/zotero/chrome.manifest"
+	echo 'pref("devtools.debugger.remote-enabled", true);' >> "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
+	echo 'pref("devtools.debugger.remote-port", 6100);' >> "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
+	echo 'pref("devtools.debugger.prompt-connection", false);' >> "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 fi
 
 echo -n "Channel: "
-grep app.update.channel "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+grep app.update.channel "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 echo
 
 # Remove unnecessary files
-${GFIND} "$BUILD_DIR" -name .DS_Store -exec rm -f {} \;
-rm -rf "$BUILD_DIR/jurism/test"
+find "$BUILD_DIR" -name .DS_Store -exec rm -f {} \;
+rm -rf "$BUILD_DIR/zotero/test"
 
 cd "$CALLDIR"
 
@@ -273,7 +273,7 @@ if [ $BUILD_MAC == 1 ]; then
 	CONTENTSDIR="$APPDIR/Contents"
 	
 	# Modify platform-specific prefs
-	perl -pi -e 's/pref\("browser\.preferences\.instantApply", false\);/pref\("browser\.preferences\.instantApply", true);/' "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+	perl -pi -e 's/pref\("browser\.preferences\.instantApply", false\);/pref\("browser\.preferences\.instantApply", true);/' "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 	
 	# Merge relevant assets from Firefox
 	mkdir "$CONTENTSDIR/MacOS"
@@ -296,7 +296,7 @@ if [ $BUILD_MAC == 1 ]; then
 	rm -f "$CONTENTSDIR/Info.plist.bak"
 	
 	# Add components
-	cp -R "$BUILD_DIR/jurism/"* "$CONTENTSDIR/Resources"
+	cp -R "$BUILD_DIR/zotero/"* "$CONTENTSDIR/Resources"
 	
 	# Add Mac-specific Standalone assets
 	cd "$CALLDIR/assets/mac"
@@ -370,7 +370,7 @@ if [ $BUILD_WIN32 == 1 ]; then
 	mkdir "$APPDIR"
 	
 	# Modify platform-specific prefs
-	perl -pi -e 's/%GECKO_VERSION%/'"$GECKO_VERSION_WIN"'/g' "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+	perl -pi -e 's/%GECKO_VERSION%/'"$GECKO_VERSION_WIN"'/g' "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 	
 	# Copy relevant assets from Firefox
 	cp -R "$WIN32_RUNTIME_PATH"/!(application.ini|browser|defaults|devtools-files|crashreporter*|firefox.exe|maintenanceservice*|precomplete|removed-files|uninstall|update*) "$APPDIR"
@@ -393,7 +393,7 @@ if [ $BUILD_WIN32 == 1 ]; then
 	cp "$CALLDIR/win/updater.exe" "$APPDIR"
 	cat "$CALLDIR/win/installer/updater_append.ini" >> "$APPDIR/updater.ini"
 	
-	cp -R "$BUILD_DIR/jurism/"* "$BUILD_DIR/application.ini" "$APPDIR"
+	cp -R "$BUILD_DIR/zotero/"* "$BUILD_DIR/application.ini" "$APPDIR"
 	
 	# Add Windows-specific Standalone assets
 	cd "$CALLDIR/assets/win"
@@ -510,7 +510,7 @@ if [ $BUILD_LINUX == 1 ]; then
 		mkdir "$APPDIR"
 		
 		# Merge relevant assets from Firefox
-		cp -r "$RUNTIME_PATH/"!(application.ini|browser|defaults|crashreporter|crashreporter.ini|firefox-bin|precomplete|removed-files|run-mozilla.sh|update-settings.ini|updater|updater.ini) "$APPDIR"
+		cp -r "$RUNTIME_PATH/"!(application.ini|browser|defaults|devtools-files|crashreporter|crashreporter.ini|firefox-bin|precomplete|removed-files|run-mozilla.sh|update-settings.ini|updater|updater.ini) "$APPDIR"
 		
 		# Use our own launcher that calls the original Firefox executable with -app
 		mv "$APPDIR"/firefox "$APPDIR"/jurism-bin
@@ -523,10 +523,10 @@ if [ $BUILD_LINUX == 1 ]; then
 		# Use our own updater, because Mozilla's requires updates signed by Mozilla
 		cp "$CALLDIR/linux/updater-$arch" "$APPDIR"/updater
 		
-		cp -R "$BUILD_DIR/jurism/"* "$BUILD_DIR/application.ini" "$APPDIR"
+		cp -R "$BUILD_DIR/zotero/"* "$BUILD_DIR/application.ini" "$APPDIR"
 		
 		# Modify platform-specific prefs
-		perl -pi -e 's/pref\("browser\.preferences\.instantApply", false\);/pref\("browser\.preferences\.instantApply", true);/' "$BUILD_DIR/jurism/defaults/preferences/prefs.js"
+		perl -pi -e 's/pref\("browser\.preferences\.instantApply", false\);/pref\("browser\.preferences\.instantApply", true);/' "$BUILD_DIR/zotero/defaults/preferences/prefs.js"
 		
 		# Add Unix-specific Standalone assets
 		cd "$CALLDIR/assets/unix"
@@ -539,38 +539,31 @@ if [ $BUILD_LINUX == 1 ]; then
 		fi
         
 		# Add word processor plug-ins
-		#mkdir "$APPDIR/extensions"
-		#cp -RH "$CALLDIR/modules/zotero-libreoffice-integration" "$APPDIR/extensions/zoteroOpenOfficeIntegration@zotero.org"
-		#perl -pi -e 's/\.SOURCE<\/em:version>/.SA.'"$VERSION"'<\/em:version>/' "$APPDIR/extensions/zoteroOpenOfficeIntegration@zotero.org/install.rdf"
-		#echo
-		#echo -n "$ext Version: "
-		#perl -ne 'print and last if s/.*<em:version>(.*)<\/em:version>.*/\1/;' "$APPDIR/extensions/zoteroOpenOfficeIntegration@zotero.org/install.rdf"
-		#echo
-		#rm -rf "$APPDIR/extensions/zoteroOpenOfficeIntegration@zotero.org/.git"
-		
-		# Add word processor plug-ins
 		mkdir "$APPDIR/extensions"
 		cp -RH "$CALLDIR/modules/jurism-libreoffice-integration" "$APPDIR/extensions/jurismOpenOfficeIntegration@juris-m.github.io"
+		perl -pi -e 's/\.SOURCE<\/em:version>/.SA.'"$VERSION"'<\/em:version>/' "$APPDIR/extensions/jurismOpenOfficeIntegration@juris-m.github.io/install.rdf"
+		echo
+		echo -n "$ext Version: "
+		perl -ne 'print and last if s/.*<em:version>(.*)<\/em:version>.*/\1/;' "$APPDIR/extensions/jurismOpenOfficeIntegration@juris-m.github.io/install.rdf"
+		echo
+		rm -rf "$APPDIR/extensions/jurismOpenOfficeIntegration@juris-m.github.io/.git"
 
         # Add Abbreviation Filter (abbrevs-filter)
-		cp -RH "$CALLDIR/modules/abbrevs-filter" "$APPDIR/extensions/abbrevs-filter@juris-m.github.io"
+		#cp -RH "$CALLDIR/modules/abbrevs-filter" "$APPDIR/extensions/abbrevs-filter@juris-m.github.io"
 
         # Add Jurisdiction Support (myles)
-		cp -RH "$CALLDIR/modules/myles" "$APPDIR/extensions/myles@juris-m.github.io"
+		#cp -RH "$CALLDIR/modules/myles" "$APPDIR/extensions/myles@juris-m.github.io"
 		
         # Add Bluebook signal helper (bluebook-signals-for-zotero)
-		cp -RH "$CALLDIR/modules/bluebook-signals-for-zotero" "$APPDIR/extensions/bluebook-signals-for-zotero@mystery-lab.com"
+		#cp -RH "$CALLDIR/modules/bluebook-signals-for-zotero" "$APPDIR/extensions/bluebook-signals-for-zotero@mystery-lab.com"
 		
         # Add ODF/RTF Scan (zotero-odf-scan)
-		cp -RH "$CALLDIR/modules/zotero-odf-scan-plugin" "$APPDIR/extensions/rtf-odf-scan-for-zotero@mystery-lab.com"
+		#cp -RH "$CALLDIR/modules/zotero-odf-scan-plugin" "$APPDIR/extensions/rtf-odf-scan-for-zotero@mystery-lab.com"
 		
 		# Delete extraneous files
 		${GFIND} "$APPDIR" -depth -type d -name .git -exec rm -rf {} \;
 		${GFIND} "$APPDIR" \( -name .DS_Store -or -name update.rdf \) -exec rm -f {} \;
 		${GFIND} "$APPDIR/extensions" -depth -type d -name build -exec rm -rf {} \;
-		
-		# Add run-zotero.sh
-		#cp "$CALLDIR/linux/run-jurism.sh" "$APPDIR/run-jurism.sh"
 		
 		if [ $PACKAGE == 1 ]; then
 			# Create tar
